Rename "router" => "routes"
